+++ src/main/java/com/example/apipeticos/repositories/RaceRepository.java
@@ -3,6 +3,6 @@ package com.example.apipeticos.repositories;
 import com.example.apipeticos.models.Race;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RaceRepository extends JpaRepository<Race, Long> {
+public interface RaceRepository extends JpaRepository<Race, Integer> {
 
 }
